Сделал рефакторинг
