+++ src/const.js
@@ -1,3 +1,10 @@
+export const RenderPosition = {
+  BEFOREBEGIN: 'beforebegin',
+  AFTERBEGIN: 'afterbegin',
+  BEFOREEND: 'beforeend',
+  AFTEREND: 'afterend',
+};
+
 export const OFFERS = [
   {
     id: '1',
